requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
tkinter
matplotlib>=3.7.0
scikit-learn>=1.3.0
dlib>=19.24.0
face-recognition>=1.3.0
albumentations>=1.3.0
seaborn>=0.12.0
scikit-image>=0.21.0
scipy>=1.11.0

# Audio processing dependencies
librosa>=0.10.0
torchaudio>=2.0.0
soundfile>=0.12.0
resampy>=0.4.0
